requirements.txt
playwright>=1.40.0
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
